currency formatter added

+++ src/model/Apartment.java
@@ -1,9 +1,11 @@
 package model;
 
-public class Apartment extends House{
+import java.math.BigDecimal;
+
+public class Apartment extends House {
     private int floor;
 
-    public Apartment(int roomCount, int livingRoomCount, double price, double squareMeter, int floor) {
+    public Apartment(int roomCount, int livingRoomCount, BigDecimal price, double squareMeter, int floor) {
         super(roomCount, livingRoomCount, price, squareMeter);
         this.floor = floor;
     }
